% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyNIS-package.R
\docType{package}
\name{easyNIS-package}
\alias{easyNIS}
\alias{easyNIS-package}
\title{Development tools for the National Inpatient Sample}
\description{
easyNIS currently provides a capability roadmap and invented fixtures.
No licensed-data year is release-supported. Import, harmonization, cohorts,
survey inference, models, and publication exports remain in development.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MSaadAsif/easyNIS}
  \item Report bugs at \url{https://github.com/MSaadAsif/easyNIS/issues}
}

}
\author{
\strong{Maintainer}: Muhammad Saad \email{msaadasif.md@gmail.com} (\href{https://orcid.org/0000-0002-1792-2357}{ORCID})

Authors:
\itemize{
  \item Muhammad Saad \email{msaadasif.md@gmail.com} (\href{https://orcid.org/0000-0002-1792-2357}{ORCID})
  \item Ali
}

}
\keyword{internal}
